Pomodoro Timer
A timer that will say work for 25mins than go to a short break for 5 mins this will repeat 4 times than go to a long break for 20mins. and repeat till reset is clicked in the app.
Was fun to figure it out learned a lot of neat things. EXE included